refactor: rename services and organize debug files
