+++ src/debug/windows/base-window.ts
