+++ src/debug/windows/peer-inspector-window.ts
@@ -1,7 +1,7 @@
 import { ImGui, ImVec2 } from "@mori2003/jsimgui";
 import { BaseWindow } from "./base-window.js";
-import { WebRTCService } from "../services/webrtc-service.js";
-import { ServiceLocator } from "../services/service-locator.js";
+import { WebRTCService } from "../../services/webrtc-service.js";
+import { ServiceLocator } from "../../services/service-locator.js";
 
 export class PeerInspectorWindow extends BaseWindow {
   private static readonly COLOR_CONNECTED_STATE = 0xff00ff00;
