+++ src/services/object-animation-service.ts
