+++ src/services/webrtc-service.ts
@@ -1,7 +1,7 @@
 import { TunnelType } from "../enums/tunnel-type.js";
 import type { WebRTCPeer } from "../interfaces/webrtc-peer.js";
 import { WebRTCPeerService } from "./webrtc-peer-service.js";
-import { DebugUtils } from "../utils/debug-utils.js";
+import { DebugUtils } from "../debug/debug-utils.js";
 import { WebSocketType } from "../enums/websocket-type.js";
 import { BinaryWriter } from "../utils/binary-writer-utils.js";
 import type { BinaryReader } from "../utils/binary-reader-utils.js";
