+++ src/utils/binary-reader-utils.ts
@@ -1,4 +1,4 @@
-import { DebugUtils } from "./debug-utils";
+import { DebugUtils } from "../debug/debug-utils.js";
 
 export class BinaryReader {
   private readonly buffer: ArrayBuffer;
