+++ src/utils/binary-writer-utils.ts
@@ -1,4 +1,4 @@
-import { DebugUtils } from "./debug-utils";
+import { DebugUtils } from "../debug/debug-utils.js";
 
 export class BinaryWriter {
   private buffer: ArrayBuffer;
